Add files via upload
This MATLAB files are the implementation of the solution for Diffraction of Elastic SH waves generated by Wedges presented in the paper:
Francisco J. Sánchez-Sesma; Diffraction of elastic SH waves by wedges. Bulletin of the Seismological Society of America 1985; 75 (5): 1435–1446. 

EXAMPLE to run/invoke the code:

wedgeSpaceWave(40, 1/2, 256, 0.04, 0.5,  0.4,  1, 0, 1.0,  -2.0,  4, 101, 101)

The parameters inside this function (which is the main code) are described inside the file wedgeSpaceWave.m
If you find this solution useful, please consider to cite the above paper.